Create Savegame object and save off credits, update credits to Int32

+++ Source/DragonRogue/Public/DaCreditsPickupItem.h
@@ -16,7 +16,7 @@ class DRAGONROGUE_API ADaCreditsPickupItem : public ADaPickupItem
 
 public:
 	UPROPERTY(EditAnywhere, Category="Credits")
-	float CreditsAmount;
+	int32 CreditsAmount;
 	
 	void Interact_Implementation(APawn* InstigatorPawn) override;
